Merge new features in main

+++ app/Http/Controllers/BuildingController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateBuildingRequest;
 use App\Models\Building;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use App\Helper\CLog;
 
 class BuildingController extends Controller
 {
@@ -19,15 +20,17 @@ class BuildingController extends Controller
     public function store(StoreBuildingRequest $request)
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'required|unique:buildings|max:100',
+            'name' => 'required|max:100',
             'location_id' => 'required|integer',
         ])->validate();
 
         if ($validator and Building::create($request->all())) {
-            LogController::log('Gebäude erstellt', '{"name": "' . $request->name . '", "location_id": "' . $request->location_id . '"}');
 
-            return redirect()->back()->with('success', __('Msg.LocationCreated'));
+            CLog::info("Building", "Create building {$request->name}");
+            return redirect()->back()->with('success', __('Msg.BuildingCreated'));
         }
+
+        CLog::error("Building", "Could not create building {$request->name}");	
         return redirect()->back()->withErrors(['message' => 'Building could not be created']);
     }
 
@@ -41,14 +44,18 @@ class BuildingController extends Controller
     public function update(UpdateBuildingRequest $request, Building $building)
     {
         $validator = Validator::make($request->all(), [
-            'name' => 'required|unique:buildings|max:100',
+            'name' => 'required|max:100',
+            'id' => 'required|integer|exists:buildings,id',
+            'location_id' => 'required|integer',
         ])->validate();
 
         if ($validator and $building->whereId($request->input('id'))->update($request->except(['_token', '_method']))) {
-            LogController::log('Gebäude aktualisiert', '{"name": "' . $request->name . '"}');
 
-            return redirect()->back()->with('success', __('Msg.LocationUpdated'));
+            CLog::info("Building", "Update building {$request->name}");
+            return redirect()->back()->with('success', __('Msg.BuildingUpdated'));
         }
+
+        CLog::error("Building", "Could not update building {$request->name}");
         return redirect()->back()->withErrors(['message' => 'Building could not be updated']);
     }
 
@@ -60,12 +67,23 @@ class BuildingController extends Controller
      */
     public function destroy(Request $building)
     {
-        $find = Building::find($building->input('id'));
-        if ($find->delete()) {
-            LogController::log('Gebäude gelöscht', '{"name": "' . $building->name . '"}');
+        $validator = Validator::make($building->all(), [
+            'id' => 'required|integer|exists:buildings,id',
+        ])->validate();
 
-            return redirect()->back()->with('success', __('Msg.LocationDeleted'));
+        $find = Building::find($building->input('id'));
+
+        if($find->rooms()->count() > 0) {
+            return redirect()->back()->withErrors(['message' => 'Building could not be deleted, because it has rooms assigned to it.']);
         }
+
+        if ($find->delete()) {
+            CLog::info("Building", "Delete building {$find->name}");
+
+            return redirect()->back()->with('success', __('Msg.BuildingDeleted'));
+        }
+
+        CLog::error("Building", "Could not delete building {$find->name}");
         return redirect()->back()->with('message', 'Could not delete building');
     }
 }
